move_dot to be done
