+++ plot_aperture_old.py
@@ -5,7 +5,7 @@ import os
 import sys
 import numpy as np
 import matplotlib.pyplot as plt
-from plot_beamline import plot_beamline
+from plot_beamline_old import plot_beamline
 
 
 def plot_aperture(cfg, ax, plane='X'):
