+++ plot_orbit_correction.py
@@ -1,428 +1,202 @@
 #!/usr/bin/env python3
 
-import json
-import os
-import sys
-import numpy as np
 import matplotlib.pyplot as plt
-from plot_beamline import plot_beamline
-from plot_aperture import plot_aperture
-from threading import Timer
-from cx_connection import CXDataExchange
+import numpy as np
+from matplotlib.patches import Rectangle
 from matplotlib.widgets import Button
-from PyQt5.QtWidgets import QApplication
+from threading import Timer
+
+from accelerator_devices import AccElemCollection
 
 
-def plot_orbit(beamline_cfg, ax, plane='X'):
-    s_values = []
-    orbit_values = []
-    BPM_names = []
+class OrbitCorMatPlot:
+    def __init__(self, get_vals_from_consys, send_vals_to_consys, **kwargs):
+        super().__init__()
+        # to be done
+        aper = kwargs.get('aper')
+        bline = kwargs.get('bline')
 
-    for name, itm in beamline_cfg["Beamline elements"].items():
-        if "type" in itm and itm["type"] == "BPM":
-            if "plane" in itm and itm["plane"] == plane:
-                if "s" in itm and "value" in itm:
-                    s_values.append(itm["s"])
-                    orbit_values.append(itm["value"])
-                    BPM_names.append(name)
-    
-    s_values = np.array(s_values)
-    orbit_values = np.array(orbit_values)
-    BPM_names = np.array(BPM_names)
-    
-    ids = np.argsort(s_values) # indicies sorted by s
-    line, = ax.plot(s_values[ids], orbit_values[ids], "--", marker="o", lw=3)
-    return line, BPM_names[ids]
+        self.bpm_col = AccElemCollection('bpm_coll', 'BPM', bline)
+        self.cor_col = AccElemCollection('cor_col', 'corrector', bline)
 
-def plot_correctors(beamline_cfg, ax, plane='X'):
-    s_values=[]
-    kick_values=[]
-    corrector_names = []
-    min_kicks = []
-    max_kicks = []
+        # matplotlib binding
+        ################################################################################################################
+        self.txt: dict = {}
+        self.fig, self.axides = plt.subplots(6, 1, sharex='col', sharey='col')
 
-    for name, itm in beamline_cfg["Beamline elements"].items():
-        if "type" in itm and itm["type"] == "corrector":
-            if "plane" in itm and itm["plane"] == plane:
-                if "s" in itm and "kick" in itm:
-                    corrector_names.append(name)
-                    s = itm["s"]
-                    s_values.append(s)
-                    kick_values.append(itm["kick"])
+        for axis in range(len(self.axides) - 2):
+            self.plot_aper(aper, self.axides[axis])
+            self.plot_beamline(bline, self.axides[axis], 10)
+            self.txt[id(self.axides[axis])] = self.axides[axis].text(0, 0, "", verticalalignment='center',
+                                                                     horizontalalignment='right', color='black',
+                                                                     fontsize=9, rotation='vertical', alpha=0.7)
 
-                    if "min_kick" in itm:
-                        min_kick = itm["min_kick"]
-                        ax.plot([s,s], [0,min_kick], color="blue", alpha=0.3, lw=3)
-                    else: min_kick = -1e10
+        btn_get = Button(self.axides[4], 'Get Values')
+        btn_get.on_clicked(get_vals_from_consys)
+        btn_set = Button(self.axides[5], 'Set Values')
+        btn_set.on_clicked(send_vals_to_consys)
 
-                    if "max_kick" in itm:
-                        max_kick = itm["max_kick"]
-                        ax.plot([s,s], [0,max_kick], color="blue", alpha=0.3, lw=3)
-                    else: max_kick = 1e10
+        self.names_id = {id(self.axides[0]): self.bpm_col.names_x, id(self.axides[1]): self.bpm_col.names_y,
+                         id(self.axides[2]): self.cor_col.names_x, id(self.axides[3]): self.cor_col.names_y}
+        self.s_id = {id(self.axides[0]): self.bpm_col.s_x, id(self.axides[1]): self.bpm_col.s_y,
+                     id(self.axides[2]): self.cor_col.s_x, id(self.axides[3]): self.cor_col.s_y}
 
-                    max_kicks.append(max_kick)
-                    min_kicks.append(min_kick)
+        self.plot_orbit()
+        self.plot_correctors()
 
-    line, = ax.plot(s_values, kick_values, " ", marker="o",
-                    markersize=8, markeredgewidth=1, markeredgecolor="black",
-                    color="white", alpha=0.5)
-    return line, corrector_names, min_kicks, max_kicks
+        self.scroll_in_progress = False
+        self.N_scrolls_done = 0
+        self.selected_name: str = ' '
+        self.selected_txt = self.txt[id(self.axides[0])]
 
-def responses_to_corrector(element_names, corrector_name, responses_cfg):
-    slopes = []
-    for name in element_names:
-        slope = 0
-        response_id = f"{name} / {corrector_name}"
-        if response_id in responses_cfg["response data"]:
-            slope = responses_cfg["response data"][response_id]["slope"]
-        slopes.append(slope)
-    return np.array(slopes)
+        self.fig.canvas.mpl_connect('scroll_event', self.on_scroll)
+        self.fig.canvas.mpl_connect('motion_notify_event', self.onmove)
 
-scroll_in_progress = False
-N_scrolls_done = 0
-def on_scroll(event):
-    global scroll_in_progress
-    global N_scrolls_done
+        self.fig.subplots_adjust(left=0.06, bottom=0.08, right=0.99, top=0.96, hspace=0)
+        plt.show()
+        ################################################################################################################
 
-    x = event.xdata
-    if not x: return
+    def plot_aper(self, aper, axis):
+        axis.plot(aper[0], aper[1], lw=2, c='black', alpha=0.4)
 
-    if event.button == "up":
-        N_scrolls_done+=1
-    else:
-        N_scrolls_done-=1
+    def plot_beamline(self, bline, axis, h):
+        color = {'BEND': 'green', 'QUAD': 'red', 'BPM': 'black', 'corrector': 'orange'}
+        for elem_name, elem_par in bline['Beamline elements'].items():
+            s = elem_par['s']
+            l = elem_par['L']
+            axis.add_patch(Rectangle((s - l/2, 0-h/2), l, h, facecolor=color[elem_par['type']], alpha=0.5))
+            # axis.plot([s, s], [0, -1], lw=1, color='blue', alpha=0.5)
+            # axis.text(s, -1, ", ".join(name), verticalalignment='bottom', horizontalalignment='right',
+            #           color='blue', fontsize=9, rotation='vertical', alpha=0.5)
 
-    if not scroll_in_progress:
-        scroll_in_progress = True
-        t = Timer(0.2, move_dot, args=[event])
-        t.start()
+    def plot_orbit(self, orbit_x=np.zeros(16,), orbit_y=np.zeros(16,)):
+        self.axides[0].plot(self.bpm_col.s_x[self.bpm_col.ind_x], orbit_x[self.bpm_col.ind_x], "--", marker="o", lw=3)
+        self.axides[1].plot(self.bpm_col.s_x[self.bpm_col.ind_y], orbit_y[self.bpm_col.ind_y], "--", marker="o", lw=3)
 
-def move_dot(event):
-    global scroll_in_progress, N_scrolls_done
-    scroll_in_progress = False
-    
-    global x_orbit_values, y_orbit_values
-    global x_cor_values, y_cor_values
-    
-    res = line_and_element_from_mouse_event(event)
-    if not res: return
-    line, i, element_name = res
+    def plot_correctors(self, x_kick=np.zeros(19,), y_kick=np.zeros(17,)):
+        self.axides[2].plot(self.cor_col.s_x[self.cor_col.ind_x], x_kick[self.cor_col.ind_x], " ", marker="o",
+                            markersize=8, markeredgewidth=1, markeredgecolor="black", color="white", alpha=0.5)
+        self.axides[3].plot(self.cor_col.s_y[self.cor_col.ind_y], y_kick[self.cor_col.ind_y], " ", marker="o",
+                            markersize=8, markeredgewidth=1, markeredgecolor="black", color="white", alpha=0.5)
 
-    v_min, v_max =event.inaxes.get_ylim()
-    step = N_scrolls_done*(v_max-v_min)/30.0
-    N_scrolls_done = 0
-    
-    element = beamline_cfg["Beamline elements"][element_name]
-    if element["type"] == "corrector":
-        if element["plane"] == "X":
-            kick_values = x_cor_values
+    def on_scroll(self, event):
+        x = event.xdata
+        if not x:
+            return
+        if event.button == "up":
+            self.N_scrolls_done += 1
         else:
-            kick_values = y_cor_values
-        
-        kick = kick_values[i]
+            self.N_scrolls_done -= 1
 
-        if "max_kick" in element:
-            if kick + step > element["max_kick"]:
-                step = element["max_kick"] - kick
-        if "min_kick" in element:
-            if kick + step < element["min_kick"]:
-                step = element["min_kick"] - kick
-        
-        x_resp = responses_to_corrector(x_BPMs, element_name, responses_cfg)
-        y_resp = responses_to_corrector(y_BPMs, element_name, responses_cfg)
-        
-        x_orbit_values = x_orbit_values + x_resp*step
-        y_orbit_values = y_orbit_values + y_resp*step
+        if not self.scroll_in_progress:
+            self.scroll_in_progress = True
+            t = Timer(0.2, self.move_dot, args=[event])
+            t.start()
 
-        kick_values[i]+= step
-        
-    elif element["type"] == "BPM":
-        dX = x_orbit_values*0.0
-        dY = y_orbit_values*0.0
-        if element["plane"] == "X":
-            dX[i] = step
-        else:
-            dY[i] = step
-        
-        dV_wish = np.concatenate( (dX,dY) )
-                
-        dkicks = ORM_inv*np.matrix(dV_wish).T
-        
-        kicks = np.concatenate((x_cor_values,y_cor_values))
-        
-        min_kicks = np.concatenate((x_min_kicks,y_min_kicks))
-        max_kicks = np.concatenate((x_max_kicks,y_max_kicks))
-        
-        new_kicks = kicks + dkicks.A1
-        
-        new_kicks = [min((kick, max_kick)) for kick, max_kick in zip(new_kicks, max_kicks)]
-        new_kicks = [max((kick, min_kick)) for kick, min_kick in zip(new_kicks, min_kicks)]
-        new_kicks = np.array(new_kicks)
+    def move_dot(self, event):
+        self.scroll_in_progress = False
 
-        dkicks = np.matrix(new_kicks-kicks).T
-        
-        dV = ORM*dkicks
-        dV = dV.A1
+        name, s = self.element_from_mouse_event(event)
+        if not name:
+            return
 
-        dkicks = dkicks.A1
-        
-        x_cor_values = x_cor_values + dkicks[:len(x_cor_values)]
-        y_cor_values = y_cor_values + dkicks[len(x_cor_values):]
-        
-        x_orbit_values = x_orbit_values + dV[:len(x_orbit_values)]
-        y_orbit_values = y_orbit_values + dV[len(x_orbit_values):]
+        v_min, v_max = event.inaxes.get_ylim()
+        step = self.N_scrolls_done * (v_max - v_min) / 30.0
+        self.N_scrolls_done = 0
 
-    update_plot()
-
-
-def get_ORM(BPMs, correctors):
-    M = []
-    for BPM in BPMs:
-        line = []    
-        for cor in correctors:
-            key = f"{BPM} / {cor}"
-            if key in responses_cfg["response data"]:
-                line.append(responses_cfg["response data"][key]["slope"])
+        element = beamline_cfg["Beamline elements"][element_name]
+        if element["type"] == "corrector":
+            if element["plane"] == "X":
+                kick_values = x_cor_values
             else:
-                line.append(0.0)
-        M.append(line)
-    
-    return np.matrix(M)
+                kick_values = y_cor_values
 
-def line_and_element_from_mouse_event(event):
-    x = event.xdata
-    if not x: return
-    
-    mouse_x = event.xdata
-    if not mouse_x or not event.inaxes: return
-    
-    if event.inaxes is axcx:
-        line = cx_dots
-        names = x_correctors
-    if event.inaxes is axcy:
-        line = cy_dots
-        names = y_correctors
-    if event.inaxes is axx:
-        line = x_line
-        names = x_BPMs
-    if event.inaxes is axy:
-        line = y_line
-        names = y_BPMs
-    
-    s_data, value_data = line.get_data()
+            kick = kick_values[i]
 
-    i = np.argmin(np.abs(s_data-mouse_x)) # nearest element index
-    element_name = names[i]
-    
-    return line, i, element_name
+            if "max_kick" in element:
+                if kick + step > element["max_kick"]:
+                    step = element["max_kick"] - kick
+            if "min_kick" in element:
+                if kick + step < element["min_kick"]:
+                    step = element["min_kick"] - kick
 
-def onmove(event):
-    res = line_and_element_from_mouse_event(event)
-    if not res: return
-    line, i, element_name = res
-    
-    select_element(element_name, event.inaxes)
+            x_resp = responses_to_corrector(x_BPMs, element_name, responses_cfg)
+            y_resp = responses_to_corrector(y_BPMs, element_name, responses_cfg)
 
-selected_element_name = ""
-selected_element_txt = None
-def select_element(name, ax):
-    global selected_element_name, selected_element_txt
-    if name == selected_element_name:
-        return
-    selected_element_name = name
-    
-    if selected_element_txt:
-        selected_element_txt.set_text("")
+            x_orbit_values = x_orbit_values + x_resp * step
+            y_orbit_values = y_orbit_values + y_resp * step
 
-    element = beamline_cfg["Beamline elements"][name]
-    s = 0
-    if "s" in element: s = element["s"]
-    
-    if ax is axcx:
-        txt = txt_cx
-    if ax is axcy:
-        txt = txt_cy
-    if ax is axx:
-        txt = txt_x
-    if ax is axy:
-        txt = txt_y
-    
-    txt.set_text(name)
-    txt.set_x(s)
-    selected_element_txt = txt
-    
-    ax.figure.canvas.draw()
+            kick_values[i] += step
 
+        elif element["type"] == "BPM":
+            dX = x_orbit_values * 0.0
+            dY = y_orbit_values * 0.0
+            if element["plane"] == "X":
+                dX[i] = step
+            else:
+                dY[i] = step
 
-def update_plot():
+            dV_wish = np.concatenate((dX, dY))
 
-    for line, values in zip([x_line, y_line, cx_dots, cy_dots],
-                            [x_orbit_values, y_orbit_values,
-                             x_cor_values, y_cor_values]):
-        
-        s_data, _ = line.get_data()
-        line.set_data(s_data, values)
+            dkicks = ORM_inv * np.matrix(dV_wish).T
 
-    fig.canvas.draw()
+            kicks = np.concatenate((x_cor_values, y_cor_values))
 
+            min_kicks = np.concatenate((x_min_kicks, y_min_kicks))
+            max_kicks = np.concatenate((x_max_kicks, y_max_kicks))
 
-def get_vals_from_consys(event):
-    x_orbit_values = []
-    y_orbit_values = []
-    coors = data_exchange.get_beam_coor()
-    for chan, coor in coors.items():
-        if chan[-1] == 'x':
-            x_orbit_values.append(coor)
-        else:
-            y_orbit_values.append(coor)
-    print(y_orbit_values)
-    cors_cur = data_exchange.get_cor_vals()
-    # update_plot()
+            new_kicks = kicks + dkicks.A1
 
+            new_kicks = [min((kick, max_kick)) for kick, max_kick in zip(new_kicks, max_kicks)]
+            new_kicks = [max((kick, min_kick)) for kick, min_kick in zip(new_kicks, min_kicks)]
+            new_kicks = np.array(new_kicks)
 
-def send_vals_to_consys(event):
-    print(event)
+            dkicks = np.matrix(new_kicks - kicks).T
 
+            dV = ORM * dkicks
+            dV = dV.A1
 
-if __name__ == '__main__':
-    app = QApplication(['resp_orbittt'])
-    data_exchange = CXDataExchange('/home/vbalakin/PycharmProjects/accel_packages/TuneBeamline/cx_data_config.json',
-                                   'bpm', 'cor')
-    beamline_file='beamline.json'
-    if len(sys.argv) > 1:
-        beamline_file = sys.argv[1]
-    if not os.path.exists(beamline_file):
-        print(f"No {beamline_file} file to plot!")
-        exit()
-    
-    responses_file='responses.json'
-    if len(sys.argv) > 2:
-        responses_file = sys.argv[2]
-    if not os.path.exists(responses_file):
-        print(f"No {responses_file} file to use!")
-        exit()
+            dkicks = dkicks.A1
 
-    aperture_file= ''  #'aperture.json'
-    if len(sys.argv) > 3:
-        aperture_file = sys.argv[3]
+            x_cor_values = x_cor_values + dkicks[:len(x_cor_values)]
+            y_cor_values = y_cor_values + dkicks[len(x_cor_values):]
 
-    with open(beamline_file, "r") as f:
-        beamline_cfg = json.load(f)
-        print(f"Beamline file: {beamline_file}")
+            x_orbit_values = x_orbit_values + dV[:len(x_orbit_values)]
+            y_orbit_values = y_orbit_values + dV[len(x_orbit_values):]
 
-    with open(responses_file, 'r') as f:
-        responses_cfg = json.load(f)
-        print(f"Response matrix file: {responses_file}")
-        
-    fig = plt.figure(figsize=(11,6))
-    
-    axx  = fig.add_subplot(411)
-    axy  = fig.add_subplot(412, sharex=axx, sharey=axx)
-    axcx = fig.add_subplot(413, sharex=axx)
-    axcy = fig.add_subplot(414, sharex=axx, sharey=axcx)
+        update_plot()
 
-    title = beamline_file
+    def onmove(self, event):
+        name, s = self.element_from_mouse_event(event)
+        if not name:
+            return
+        self.select_element(name, s, event.inaxes)
 
-    axx.set_title(title)
+    def element_from_mouse_event(self, event):
+        x = event.xdata
+        if not x:
+            return False, False
+        mouse_x = event.xdata
+        if not mouse_x or not event.inaxes:
+            return False, False
+        try:
+            i = np.argmin(np.abs(self.s_id[id(event.inaxes)] - mouse_x))  # nearest element index
+        except KeyError:
+            return False, False
 
-    if os.path.exists(aperture_file):
-        with open(aperture_file, "r") as f:
-            aperture_cfg = json.load(f)
+        return self.names_id[id(event.inaxes)][i], self.s_id[id(event.inaxes)][i]
 
-            plot_aperture(aperture_cfg, axx, plane="X")
-            plot_aperture(aperture_cfg, axy, plane="Y")
+    def select_element(self, name, s, ax):
+        if name == self.selected_name:
+            return
+        self.selected_name = name
 
-    plot_beamline(beamline_cfg, axx,  show_names=False, box_h=10.0, alpha=0.2)
-    plot_beamline(beamline_cfg, axy,  show_names=False, box_h=10.0, alpha=0.2)
-    plot_beamline(beamline_cfg, axcx, show_names=False, box_h=1.0, alpha=0.2)
-    plot_beamline(beamline_cfg, axcy, show_names=False, box_h=1.0, alpha=0.2)
+        if self.selected_txt:
+            self.selected_txt.set_text("")
 
-    x_line, x_BPMs = plot_orbit(beamline_cfg, axx, plane='X')
-    y_line, y_BPMs = plot_orbit(beamline_cfg, axy, plane='Y')
+        txt = self.txt[id(ax)]
+        txt.set_text(name)
+        txt.set_x(s)
+        self.selected_txt = txt
 
-    txt_x = axx.text(0, 0, "", verticalalignment='center',
-                     horizontalalignment='right', color='black',
-                     fontsize=9, rotation='vertical', alpha=0.7)
-    txt_y = axy.text(0, 0, "", verticalalignment='center',
-                     horizontalalignment='right', color='black',
-                     fontsize=9, rotation='vertical', alpha=0.7)
-    txt_cx = axcx.text(0, 0, "", verticalalignment='center',
-                     horizontalalignment='right', color='black',
-                     fontsize=9, rotation='vertical', alpha=0.7)
-    txt_cy = axcy.text(0, 0, "", verticalalignment='center',
-                     horizontalalignment='right', color='black',
-                     fontsize=9, rotation='vertical', alpha=0.7)
-
-    x_orbit_values = x_line.get_data()[1] # mm
-    y_orbit_values = y_line.get_data()[1] # mm
-
-    cx_dots, x_correctors, x_min_kicks, x_max_kicks = \
-        plot_correctors(beamline_cfg, axcx, plane='X')
-    cy_dots, y_correctors, y_min_kicks, y_max_kicks = \
-        plot_correctors(beamline_cfg, axcy, plane='Y')
-    
-    x_cor_values = cx_dots.get_data()[1] # mrad
-    y_cor_values = cy_dots.get_data()[1] # mrad
-    
-    #print(x_line.get_data())
-
-    if len(x_correctors) > 0:
-        cor = x_correctors[0]
-    elif len(y_correctors) > 0:
-        cor = y_correctors[0]
-    kick_units = beamline_cfg["Beamline elements"][cor]["kick_units"]
-
-    axx.set_ylabel("$x$ (mm)")
-
-    axy.set_ylabel("$y$ (mm)")
-
-    axcx.set_ylabel(f"x-corr. ({kick_units})")
-    axcy.set_ylabel(f"y-corr. ({kick_units})")
-    axcy.set_xlabel("$s$ (m)")
-
-    fig.subplots_adjust(left=0.06, bottom=0.08, right=0.99, top=0.96, hspace=0)
-    #cid = fig.canvas.mpl_connect('scroll_event', on_scroll)
-    cid = fig.canvas.mpl_connect('scroll_event', on_scroll)
-    cid = fig.canvas.mpl_connect('motion_notify_event', onmove)
-
-    # btn for control
-    fig2 = plt.figure(figsize=(11, 6))
-    axget = fig2.add_subplot(211)
-    btn_get = Button(axget, 'Get Values')
-    btn_get.on_clicked(get_vals_from_consys)
-
-    axset = fig2.add_subplot(212)
-    btn_set = Button(axset, 'Set Values')
-    btn_set.on_clicked(send_vals_to_consys)
-
-    # ORM/SVD calculations:
-    
-    ORM_BPMs = np.concatenate( (x_BPMs,y_BPMs) ) # all BPMs
-    ORM_cors = np.concatenate( (x_correctors,y_correctors) ) # all correctors
-    ORM = get_ORM(ORM_BPMs,ORM_cors) # ORM matrix
-    
-    print("ORM:")
-    print(np.round(ORM, 1))
-    
-    U, s, Vh = np.linalg.svd(ORM, full_matrices=False)
-    print("Non-zero singular values:")
-    s_limit = 1e-5
-    print(s[s>s_limit])
-   
-    N_singular_values_to_keep = int( 0.85 * len(s[s>s_limit]) )
-    #N_singular_values_to_keep = 7
-    print(f"N_singular_values_to_keep = {N_singular_values_to_keep} ({100*N_singular_values_to_keep/len(s):.0f}%)")
-    s1 = s.copy()
-    s1 = s1**-1
-    s1[N_singular_values_to_keep:] = 0
-    ORM_reduced = np.matrix(U)*np.matrix(np.diag(s1))*np.matrix(Vh)  
-    #print("ORM_reduced:")
-    #print(np.round(ORM_reduced))
-
-    ORM_inv=Vh.T*np.diag(s1)*U.T
-    
-    print("Inverted ORM:")
-    print(np.round(ORM_inv))
-
-    plt.show()
-    sys.exit(app.exec_())
+        ax.figure.canvas.draw()
